+++ Game/AI/Decks/BlueEyesExecutor.cs
@@ -133,7 +133,7 @@ namespace WindBot.Game.AI.Decks
             SoulChargeUsed = false;
         }
 
-        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, int hint, bool cancelable)
+        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, long hint, bool cancelable)
         {
             Logger.DebugWriteLine("OnSelectCard " + cards.Count + " " + min + " " + max);
             if (max == 2 && cards[0].Location == CardLocation.Deck)
