+++ Game/AI/Decks/DragunExecutor.cs
@@ -136,7 +136,7 @@ namespace WindBot.Game.AI.Decks
             return 0;
         }
 
-        public override int OnSelectPlace(int cardId, int player, CardLocation location, int available)
+        public override int OnSelectPlace(long cardId, int player, CardLocation location, int available)
         {
             if (location == CardLocation.MonsterZone)
             {
